+++ src/gamspy/_algebra/condition.py
@@ -61,4 +61,7 @@ class Condition:
             self._symbol.container[self._symbol.parent.name]._is_dirty = True
             self._symbol.parent._assignment = statement
 
+        if isinstance(self._symbol, implicits.ImplicitEquation):
+            self._symbol.parent._definition = statement
+
         self._symbol.container._run()