+++ src/gamspy/_container.py
@@ -983,8 +983,8 @@ class Container(gt.Container):
 
         # if already defined equations exist, add them to .gms file
         for equation in self.getEquations():
-            if equation._assignment is not None:
-                m._add_statement(equation._assignment)
+            if equation._definition is not None:
+                m._add_statement(equation._definition)
 
         return m