+++ tests/integration/test_solve.py
@@ -1061,15 +1061,15 @@ class SolveSuite(unittest.TestCase):
         f = Parameter(m, "f")
         f[...] = 5
         
-        self.assertEqual(f.getDefinition(), "f = 5;")
+        self.assertEqual(f.getAssignment(), "f = 5;")
             
     def test_variable_discovery(self):
         x = self.m.addVariable('x')
         l2 = self.m.addModel('l2', problem=Problem.QCP, sense=Sense.MIN, objective=sqr(x - 1) + sqr(x-2))
-        self.assertTrue('x' in l2.equations[0]._assignment._find_variables())
+        self.assertTrue('x' in l2.equations[0]._definition._find_variables())
         
         e = Equation(self.m, "e", definition=(x+5) == 0)
-        self.assertTrue('x' in e._assignment._find_variables())
+        self.assertTrue('x' in e._definition._find_variables())
         
     def test_invalid_arguments(self):
         m = Container(