+++ CHANGELOG.md
@@ -8,6 +8,7 @@ GAMSPy 0.12.5
   - Add extrinsic function support.
   - Expose addGamsCode to user.
   - Refactor the underlying implementation of options.
+  - Show better error messages
 - Testing
   - Add tests for extrinsic functions.
   - Test whether the given solver is capable of solving the problem type.